move zen mode in its own file
