+++ src/vs/workbench/browser/actions/layoutActions.ts
@@ -10,10 +10,10 @@ import { IConfigurationService } from '../../../platform/configuration/common/co
 import { alert } from '../../../base/browser/ui/aria/aria.js';
 import { EditorActionsLocation, EditorTabsMode, IWorkbenchLayoutService, LayoutSettings, Parts, Position, ZenModeSettings, positionToString } from '../../services/layout/browser/layoutService.js';
 import { ServicesAccessor, IInstantiationService } from '../../../platform/instantiation/common/instantiation.js';
-import { KeyMod, KeyCode, KeyChord } from '../../../base/common/keyCodes.js';
+import { KeyMod, KeyCode } from '../../../base/common/keyCodes.js';
 import { isWindows, isLinux, isWeb, isMacintosh, isNative } from '../../../base/common/platform.js';
 import { IsMacNativeContext } from '../../../platform/contextkey/common/contextkeys.js';
-import { KeybindingsRegistry, KeybindingWeight } from '../../../platform/keybinding/common/keybindingsRegistry.js';
+import { KeybindingWeight } from '../../../platform/keybinding/common/keybindingsRegistry.js';
 import { ContextKeyExpr, ContextKeyExpression, IContextKeyService } from '../../../platform/contextkey/common/contextkey.js';
 import { IViewDescriptorService, ViewContainerLocation, IViewDescriptor, ViewContainerLocationToString } from '../../common/views.js';
 import { IViewsService } from '../../services/views/common/viewsService.js';
@@ -433,7 +433,7 @@ registerAction2(ToggleStatusbarVisibilityAction);
 
 // ------------------- Editor Tabs Layout --------------------------------
 
-abstract class AbstractSetShowTabsAction extends Action2 {
+export abstract class AbstractSetShowTabsAction extends Action2 {
 
 	constructor(private readonly settingName: string, private readonly value: string, title: ICommandActionTitle, id: string, precondition: ContextKeyExpression, description: string | ILocalizedString | undefined) {
 		super({
@@ -465,6 +465,8 @@ export class HideEditorTabsAction extends AbstractSetShowTabsAction {
 	}
 }
 
+// --- Hide Editor Tabs (Zen Mode)
+
 export class ZenHideEditorTabsAction extends AbstractSetShowTabsAction {
 
 	static readonly ID = 'workbench.action.zenHideEditorTabs';
@@ -489,6 +491,8 @@ export class ShowMultipleEditorTabsAction extends AbstractSetShowTabsAction {
 		super(LayoutSettings.EDITOR_TABS_MODE, EditorTabsMode.MULTIPLE, title, ShowMultipleEditorTabsAction.ID, precondition, localize2('showMultipleEditorTabsDescription', "Show Tab Bar with multiple tabs"));
 	}
 }
+
+// --- Show Multiple Editor Tabs (Zen Mode)
 
 export class ZenShowMultipleEditorTabsAction extends AbstractSetShowTabsAction {
 
@@ -516,6 +520,12 @@ export class ShowSingleEditorTabAction extends AbstractSetShowTabsAction {
 	}
 }
 
+registerAction2(HideEditorTabsAction);
+registerAction2(ShowMultipleEditorTabsAction);
+registerAction2(ShowSingleEditorTabAction);
+
+// --- Show Single Editor Tab (Zen Mode)
+
 export class ZenShowSingleEditorTabAction extends AbstractSetShowTabsAction {
 
 	static readonly ID = 'workbench.action.zenShowEditorTab';
@@ -528,13 +538,6 @@ export class ZenShowSingleEditorTabAction extends AbstractSetShowTabsAction {
 	}
 }
 
-registerAction2(HideEditorTabsAction);
-registerAction2(ZenHideEditorTabsAction);
-registerAction2(ShowMultipleEditorTabsAction);
-registerAction2(ZenShowMultipleEditorTabsAction);
-registerAction2(ShowSingleEditorTabAction);
-registerAction2(ZenShowSingleEditorTabAction);
-
 // --- Tab Bar Submenu in View Appearance Menu
 
 MenuRegistry.appendMenuItem(MenuId.MenubarAppearanceMenu, {
@@ -543,14 +546,6 @@ MenuRegistry.appendMenuItem(MenuId.MenubarAppearanceMenu, {
 	group: '3_workbench_layout_move',
 	order: 10,
 	when: ContextKeyExpr.and(InEditorZenModeContext.negate(), IsSessionsWindowContext.negate())
-});
-
-MenuRegistry.appendMenuItem(MenuId.MenubarAppearanceMenu, {
-	submenu: MenuId.EditorTabsBarShowTabsZenModeSubmenu,
-	title: localize('tabBar', "Tab Bar"),
-	group: '3_workbench_layout_move',
-	order: 10,
-	when: ContextKeyExpr.and(InEditorZenModeContext, IsSessionsWindowContext.negate())
 });
 
 // --- Show Editor Actions in Title Bar
@@ -728,53 +723,6 @@ registerAction2(class extends Action2 {
 
 		return configurationService.updateValue('workbench.editor.pinnedTabsOnSeparateRow', newSettingValue);
 	}
-});
-
-// --- Toggle Zen Mode
-
-registerAction2(class extends Action2 {
-
-	constructor() {
-		super({
-			id: 'workbench.action.toggleZenMode',
-			title: {
-				...localize2('toggleZenMode', "Toggle Zen Mode"),
-				mnemonicTitle: localize({ key: 'miToggleZenMode', comment: ['&& denotes a mnemonic'] }, "Zen Mode"),
-			},
-			precondition: ContextKeyExpr.and(IsAuxiliaryWindowFocusedContext.toNegated(), IsSessionsWindowContext.negate()),
-			category: Categories.View,
-			f1: true,
-			keybinding: {
-				weight: KeybindingWeight.WorkbenchContrib,
-				primary: KeyChord(KeyMod.CtrlCmd | KeyCode.KeyK, KeyCode.KeyZ)
-			},
-			toggled: InEditorZenModeContext,
-			menu: [{
-				id: MenuId.MenubarAppearanceMenu,
-				group: '1_toggle_view',
-				order: 2,
-				when: IsSessionsWindowContext.negate()
-			}]
-		});
-	}
-
-	run(accessor: ServicesAccessor): void {
-		return accessor.get(IWorkbenchLayoutService).toggleZenMode();
-	}
-});
-
-KeybindingsRegistry.registerCommandAndKeybindingRule({
-	id: 'workbench.action.exitZenMode',
-	weight: KeybindingWeight.EditorContrib - 1000,
-	handler(accessor: ServicesAccessor) {
-		const layoutService = accessor.get(IWorkbenchLayoutService);
-		const contextKeyService = accessor.get(IContextKeyService);
-		if (InEditorZenModeContext.getValue(contextKeyService)) {
-			layoutService.toggleZenMode();
-		}
-	},
-	when: InEditorZenModeContext,
-	primary: KeyChord(KeyCode.Escape, KeyCode.Escape)
 });
 
 // --- Toggle Menu Bar
